+++ lib/rubex/ast.rb
@@ -1,6 +1,7 @@
-require 'rubex/ast/node'
+require 'rubex/helpers'
 require 'rubex/ast/statement'
 Dir['./lib/rubex/ast/statement/**/*.rb'].sort.each { |f| require f }
 require 'rubex/ast/expression'
 Dir['./lib/rubex/ast/expression/**/*.rb'].sort.each { |f| require f }
 require 'rubex/ast/top_statement'
+require 'rubex/ast/node'
